記事を3件まで取得して表示する
